Add marketing component to billing gateway
The pricing table component has been moved to a new 'marketing.tsx' file. This structure change simplifies the import statements as the 'marketing.tsx' is now exported in the 'package.json' of the billing gateway. Necessary import adjustments have been made in the app where this component is being used.